+++ judge-stack/src/main/java/com/dwarfeng/judge/stack/service/JudgeQosService.java
@@ -1,5 +1,7 @@
 package com.dwarfeng.judge.stack.service;
 
+import com.dwarfeng.judge.stack.bean.JudgeInfo;
+import com.dwarfeng.subgrade.stack.bean.key.LongIdKey;
 import com.dwarfeng.subgrade.stack.exception.ServiceException;
 import com.dwarfeng.subgrade.stack.service.Service;
 
@@ -16,14 +18,23 @@ public interface JudgeQosService extends Service {
      *
      * @throws ServiceException 服务异常。
      */
-    void startJudge() throws ServiceException;
+    void start() throws ServiceException;
 
     /**
      * 停止判断服务。
      *
      * @throws ServiceException 服务异常。
      */
-    void stopJudge() throws ServiceException;
+    void stop() throws ServiceException;
+
+    /**
+     * 获取指定部件的判断信息。
+     *
+     * @param sectionKey 指定的部件。
+     * @return 指定的部件对应判断信息，或者是null。
+     * @throws ServiceException 服务异常。
+     */
+    JudgeInfo getContext(LongIdKey sectionKey) throws ServiceException;
 
     /**
      * 清除本地缓存。